fix: retries pushing new image file
This is an attempt to fix the problem of multiple update messages coming
it at the same time. Usually this is from a repository that needs to use
the same image in mulitple namespaces and so needs to update mulitple
image files.

This is done by retrying the pull and push step up to five times.